windows service port from LOG-10066_MSI branch, added logger to file
